修复定位失败和步频诊断

+++ app/src/main/java/moe/fuqiuluo/portal/service/MockServiceHelper.kt
@@ -242,7 +242,7 @@ object MockServiceHelper {
         val rely = Bundle()
         rely.putString("command_id", "get_speed")
         if(locationManager.sendExtraCommand(PROVIDER_NAME, randomKey, rely)) {
-            return rely.getFloat("speed")
+            return rely.getDouble("speed").toFloat()
         }
         return null
     }
